rationai/imreg/requirements.txt
# histomicstk==1.1.0
#   Install manually: $ python -m pip install histomicstk --find-links https://girder.github.io/large_image_wheels
memory-profiler==0.58.0
numpy==1.18.5
openslide-python==1.1.2
scipy==1.6.1
scikit-image==0.18.1
Shapely==1.7.1
scikit-learn==0.24.1
tqdm==4.49.0
